property crud completed

+++ server/src/utils/cloudinary.js
@@ -1,25 +1,27 @@
 import {v2 as cloudinary} from 'cloudinary';
-import fs from "fs";
+import fs from 'fs';
 
 cloudinary.config({ 
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-  api_key: process.env.CLOUDINARY_API_KEY , 
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+  api_key: process.env.CLOUDINARY_API_KEY, 
   api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
-const uploadFileOnCloudinary = async(localFilePath)=>{
-    try{
-        const response = cloudinary.uploader.upload(localFilePath,{
-            resource_type: "auto"
-        });
-        
-        fs.unlinkSync(localFilePath);
+const uploadOnCloudinary = async(localFilePath)=>{
+  try{
+    if(!localFilePath) return null;
 
-        return response;
+    const response = await cloudinary.uploader.upload(localFilePath, {
+      resource_type: "auto"
+    })
+    
+    fs.unlinkSync(localFilePath);
 
-    }catch(error){
-        fs.unlinkSync(localFilePath);
-    }
+    return response;
+
+  }catch(error){
+    fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
+  }
 }
 
-export {uploadFileOnCloudinary};
+export {uploadOnCloudinary}
